minor mods

+++ admin/adminincludes/view_all_posts.php
@@ -130,7 +130,7 @@ if(isset($_POST['checkBoxArray'])){
                                     echo "<td>$post_tags</td>";                                  
                                     echo "<td>$post_date</td>";
                                     echo "<td><a href='post.php?source=edit_post&p_id={$post_id}'>EDIT</a></td>";
-                                    echo "<td><a href='post.php?delete={$post_id}'>DELETE</a></td>";
+                                    echo "<td><a onClick=\"javascript: return confirm('Are you Sure?');\" href='post.php?delete={$post_id}'>DELETE</a></td>";
                                     
 
                                     //echo "<td>$post_comment_count</td>";
